Fix some EPBuffer read issues
I admit, I have absolutely no idea what this does, it's a change from @bjc,
which fixes both the data loss / corruption issues we've been seeing on the
Model 100 when used with Chrysalis, and it also fixes the case where the
communication completely stalled when Chrysalis sent more than ~64 bytes at a
time.

Once we have this upstreamed and explained, we'll want to replace this commit
with the final one.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>