reset EPBuffer on SetConfiguration
This prevents a somewhat rare lockup situation. If the Model 100 is
connected to a host via a powered hub, and the hub is disconnected from
the host, the downstream ports go into suspend. If the Model 100
attempts to send two HID reports while the hub is disconnected,
and the hub is later reconnected, the Model 100 will lock up on the next
attempt to send an HID report. This appears to occur because the bus
reset that occurs when the hub is reconnected puts the `EPBuffer` for
the HID interface into an inconsistent state.

If `EPBuffer::flush()` is waiting for a transmission to complete, it can
get a spurious indication of completion after a bus reset, which sets
`txWaiting = false`. It will then queue another write on the USBD
peripheral, and set `txWaiting = true`. However, a SetConfiguration
request can reset the endpoint in the peripheral without changing
`txWaiting`.

Make an additional call to `EPBuffer::init()` in `ClassCore::init()` to
reset `txWaiting` to ensure that the `EPBuffer` state is consistent with
the endpoint state in the peripheral.

Signed-off-by: Taylor Yu <code@argon.blue>