work around udev race with 1200bps touch
On Linux systems using udev, it's possible for `dfu-util` to attempt
the upload before udev has changed the permissions to allow a non-root
user to access the DFU device. Work around this race condition by
telling Arduino to wait for the upload port, which will never appear,
because Arduino expects a serial port device. However, Arduino will
time out and fall back to running the upload command anyway. This
produces an error message, but the flashing will succeed.

Thanks to @Lilith-Daemon and @dakkar for help with troubleshooting and
testing solutions.

Signed-off-by: Taylor Yu <code@argon.blue>