NAK instead of ACK when clearing stall
When clearing an endpoint halt condition, set endpoint status to NAK
instead of VALID.

When Windows receives more data than expected from an interrupt
endpoint, it resets the device and attempts to clear the endpoint halt
condition on that endpoint. On an IN endpoint, setting the status to
VALID will cause a retransmission of the offending packet, resulting in
a loop of attempted recovery operations.

Signed-off-by: Taylor Yu <code@argon.blue>