use low-level API for sendControl
Use the low-level USBD API for USBCore_::sendControl. This has the
limitation of using a fixed-size static buffer for the total size of a
control transfer. An application can redefine the fixed limit at compile
time.

This is part of a series of changes that will eventually allow us to
remove a large amount of local duplication of the vendor ISR code.
That kind of local code duplication is likely to be risky and
introduce race conditions, no matter how carefully it's implemented.

Signed-off-by: Taylor Yu <code@argon.blue>