pre-build USB config descriptors
Construct the USB configuration descriptor set on USB reset and save
it for later, instead of constructing it each time there is a request.

This allows us to use the vendor firmware to respond to almost all of
the standard USB requests, which means we can remove a lot of code that
duplicated large parts of the vendor firmware.

Signed-off-by: Taylor Yu <code@argon.blue>