delay clearing EP_RX_ST
Avoid a race condition by delaying clearing EP_RX_ST on EP0 until the
packet buffer has been read.

The RX_ST flag "freezes" the SETUP flag. (This is not officially
documented.) If RX_ST is cleared, an incoming OUT packet can apparently
reset the SETUP flag, even if the peripheral answers it with a NAK.
This can cause the SETUP packet to be processed as an OUT packet,
resulting in application misbehavior.

Signed-off-by: Taylor Yu <code@argon.blue>