USBD low-level fix for string desc index
Use USB_STRING_COUNT instead of STR_IDX_MAX to bounds-check the
requests for string descriptors. This seems to be the intent of
the API, because the vendor examples set USB_STRING_COUNT in the
application-specific file usbd_conf.h.

In contrast, STR_IDX_MAX is hard-coded in the low-level firmware's
usbd_enum.h, and might be too high for some applications.

Signed-off-by: Taylor Yu <code@argon.blue>