use low-level API for control writes
Use more of the low-level USBD API to handle control writes. This
depends on a bugfix to the low-level firmware. Defer the validation of
the control request setup header until the Data OUT stage has
completed. This allows the recvControl to be called as expected from
within the Arduino API setup handlers of the class-specific drivers.
There is a fixed-size static buffer to receive the Data OUT payload,
shared with the Data IN payload.

This is part of a series of changes that will eventually allow us to
remove a large amount of local duplication of the vendor ISR code.
That kind of local code duplication is likely to be risky and
introduce race conditions, no matter how carefully it's implemented.

Signed-off-by: Taylor Yu <code@argon.blue>