fix some I2C hangs
Reset the I2C peripheral on a BUSY condition, because it will probably
not clear up otherwise. This was causing permanent lockups on bus
glitches.

Rework a bunch of the timeout loops to exit early when detecting an
error, instead of waiting for the full timeout period. This speeds up a
lot of common cases, such as a NACK due to a missing target.

Decrease a lot of the timeouts. The previous default was effectively
over 500ms long, which is kind of excessive for a keyboard with a 1ms
polling interval.

Signed-off-by: Taylor Yu <code@argon.blue>