correctly reset endpoints on SetConfiguration
USB 2.0 requires that the data toggles on endpoints be reset after any
configuration event, which includes a SetConfiguration request that sets
the same configuration value.

This prevents a possible problem where a host makes a SetConfiguration
request without a preceding reset, which could cause data toggles to
get out of sync, which could cause the host to ignore packets.

Signed-off-by: Taylor Yu <code@argon.blue>